+++ internal/dragonchain/prepare.go
@@ -5,23 +5,36 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
+	"strconv"
+	"time"
 
 	"github.com/dragonchain/dragonchain-installer/internal/configuration"
+	"github.com/dragonchain/dragonchain-installer/internal/helm"
 )
 
-func doesHelmDeploymentExist(name string) (bool, error) {
-	cmd := exec.Command("helm", "list", name, "--kube-context", configuration.MinikubeContext)
-	cmd.Stderr = os.Stderr
-	output, err := cmd.Output()
+func doesHelmDeploymentExist(name string, namespace string) (bool, error) {
+	helmVersion, err := helm.GetHelmMajorVersion()
 	if err != nil {
-		return false, errors.New("Error checking helm for " + name + ":\n" + err.Error())
+		return false, err
 	}
-	// If there is output, a helm installation already exists for this deployment
-	return len(output) > 0, nil
+	cmd := exec.Command("helm", "get", "notes", name, "--kube-context", configuration.MinikubeContext)
+	if helmVersion > 2 {
+		cmd = exec.Command("helm", "get", "notes", name, "-n", namespace, "--kube-context", configuration.MinikubeContext)
+	}
+	cmd.Stderr = os.Stderr
+	if err := cmd.Run(); err != nil {
+		// Will get an error if the deployment does not exist
+		return false, nil
+	}
+	// If command ran successfully, a helm installation already exists for this deployment
+	return true, nil
 }
 
 // SetupDragonchainPreReqs sets up kubernetes resource requirements for dragonchain
 func SetupDragonchainPreReqs(config *configuration.Configuration) error {
+	if err := exec.Command("kubectl", "apply", "-f", "https://raw.githubusercontent.com/rancher/local-path-provisioner/master/deploy/local-path-storage.yaml").Run(); err != nil {
+		return errors.New("Error creating local path provisioner:\n" + err.Error())
+	}
 	if exec.Command("kubectl", "get", "namespace", "dragonchain", "--context="+configuration.MinikubeContext).Run() != nil {
 		// Create the dragonchain namespace if necessary
 		fmt.Println("Creating dragonchain namespace")
@@ -34,7 +47,7 @@ func SetupDragonchainPreReqs(config *configuration.Configuration) error {
 	// Set up l1 dependencies if needed
 	if config.Level == 1 {
 		// Set up openfaas
-		exists, err := doesHelmDeploymentExist("openfaas")
+		exists, err := doesHelmDeploymentExist("openfaas", "openfaas")
 		if err != nil {
 			return errors.New("Error checking for existing openfaas installation:\n" + err.Error())
 		}
@@ -44,8 +57,30 @@ func SetupDragonchainPreReqs(config *configuration.Configuration) error {
 				return err
 			}
 		}
+		if !config.UseVM {
+			// Try to backup old docker daemon config if it exists
+			cmd := exec.Command("sudo", "mv", "/etc/docker/daemon.json", "/etc/docker/daemon.json.bak")
+			cmd.Stdin = os.Stdin
+			cmd.Run()
+			// If using native machine docker, need to ensure that insecure registry for the registry is set on the daemon
+			dockerDaemonJSON := "{\\\"insecure-registries\\\":[\\\"" + configuration.RegistryIP + ":" + strconv.Itoa(configuration.RegistryPort) + "\\\"]}"
+			cmd = exec.Command("sh", "-c", "echo "+dockerDaemonJSON+" | sudo tee /etc/docker/daemon.json")
+			cmd.Stderr = os.Stderr
+			cmd.Stdin = os.Stdin
+			if err := cmd.Run(); err != nil {
+				return errors.New("Error setting insecure registry setting with docker daemon:\n" + err.Error())
+			}
+			cmd = exec.Command("sudo", "service", "docker", "restart")
+			cmd.Stderr = os.Stderr
+			cmd.Stdin = os.Stdin
+			if err := cmd.Run(); err != nil {
+				return errors.New("Error restarting docker daemon:\n" + err.Error())
+			}
+			// Briefly wait for containers to come back up after restarting
+			time.Sleep(10 * time.Second)
+		}
 		// Set up docker registry
-		exists, err = doesHelmDeploymentExist("registry")
+		exists, err = doesHelmDeploymentExist("registry", "registry")
 		if err != nil {
 			return errors.New("Error checking for existing container registry installation:\n" + err.Error())
 		}